Put hooks into query folder
