+++ client/src/components/Gene/Categories/index.ts
