+++ client/src/hooks/queries/useGetCategories.ts
